Rename DefaultStates to PresetStates
