+++ frontend/src/Interfaces/Actions/ProjectConfigActions.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@visdesignlab/trrack';
-import { defaultState } from '../DefaultStates/DefaultState';
+import { defaultState } from '../PresetStates/DefaultState';
 import { ActionEvents } from '../Types/EventTypes';
 import { LayoutElement } from '../Types/LayoutTypes';
 import { ApplicationState } from '../Types/StateTypes';
